+++ internal/provider/system/security_ssrf_protection_resource.go
@@ -117,7 +117,7 @@ func (r *securitySsrfProtectionResource) ImportState(ctx context.Context, req re
 // Create creates the resource and sets the initial Terraform state.
 func (r *securitySsrfProtectionResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
 	var plan model.SsrfProtectionModel
-	resp.Diagnostics.Append(req.Config.Get(ctx, &plan)...)
+	resp.Diagnostics.Append(req.Plan.Get(ctx, &plan)...)
 
 	if resp.Diagnostics.HasError() {
 		tflog.Error(ctx, fmt.Sprintf("Getting request data has errors: %v", resp.Diagnostics.Errors()))