+++ internal/provider/task/task_repository_x_resource.go
@@ -42,3 +42,10 @@ func NewTaskRepositoryMavenRemoveSnapshotsResource() resource.Resource {
 		TaskType: tasktype.NewRepositoryMavenRemoveSnapshotsTask(),
 	}
 }
+
+// NewTaskRepositoryPurgeUnusedResource is a helper function to simplify the provider implementation.
+func NewTaskRepositoryPurgeUnusedResource() resource.Resource {
+	return &taskResource{
+		TaskType: tasktype.NewRepositoryPurgeUnusedTask(),
+	}
+}